updated use case scripts
