+++ src/test/K6/src/tests/platform/storage/deleteinstances.js
@@ -1,20 +1,20 @@
 /*
     Test data required: username and password, deployed app that requires level 2 login (reference app: ttd/apps-test)
-    Command: docker-compose run k6 run /src/tests/platform/storage/deleteinstances.js
-    -e env=*** -e org=*** -e username=*** -e userpwd=*** -e level2app=*** -e appnames=app1,app2 -e appsaccesskey=*** -e sblaccesskey=***
+    Command: docker-compose run k6 run /src/tests/platform/storage/deleteinstances.js -e env=*** -e org=*** -e username=*** -e userpwd=*** -e level2app=*** -e appIds=org/app1,org/app2 -e appsaccesskey=*** -e sblaccesskey=***
 */
 
 import { check } from 'k6';
 import * as sbl from '../../../api/platform/storage/messageboxinstances.js';
 import * as setUpData from '../../../setup.js';
 import { addErrorCount } from '../../../errorcounter.js';
+import * as instances from '../../../api/platform/storage/instances.js';
 
 const userName = __ENV.username;
 const userPassword = __ENV.userpwd;
 const appOwner = __ENV.org;
 const level2App = __ENV.level2app;
 const environment = __ENV.env.toLowerCase();
-let appNames = __ENV.appnames;
+let appIds = __ENV.appIds;
 
 export const options = {
   thresholds: {
@@ -32,7 +32,7 @@ export function setup() {
   return data;
 }
 
-//Hard delete instances under a party id based on supplied app names
+//Hard delete instances under a party id based on supplied appIds + level2app
 export default function (data) {
   const runtimeToken = data['RuntimeToken'];
 
@@ -41,42 +41,44 @@ export default function (data) {
   //add party id of org for non prod environments
   if (environment != 'prod') partyIds.push(data['orgNumberPartyId']);
 
-  var res, success, instances;
+  var res, resBody, success, instanceIds;
   var instancesCount = 0;
 
   try {
-    appNames = appNames.split(',');
-    appNames.push(level2App);
+    appIds = appIds.split(',');
+    appIds.push(appOwner + '/' + level2App);
   } catch (error) {
-    appNames = [];
-    appNames.push(level2App);
+    appIds = [];
+    appIds.push(appOwner + '/' + level2App);
   }
+
   partyIds.forEach((partyId) => {
+    //Find active instances under the party id to be deleted.
+    var filters = {
+      'instanceOwner.partyId': partyId,
+      appId: appIds,
+    };
+
+    res = instances.getAllinstancesWithFilters(runtimeToken, filters);
+    resBody = JSON.parse(res.body);
+    instanceIds = resBody.instances.map((i) => i.id);
+    instancesCount = instanceIds.count;
+
     do {
-      //Find active instances under the party id to be deleted.
-      var filters = {
-        'instanceOwner.partyId': partyId,
-      };
-      res = sbl.searchSblInstances(runtimeToken, filters);
-
-      //Filter instances based on appName
-      instances = sbl.filterInstancesByAppName(appNames, res.body);
-      instancesCount = instances.length;
-
-      //hard delete all the instances fetched
       if (instancesCount > 0) {
-        sbl.hardDeleteManyInstances(runtimeToken, instances);
+        sbl.hardDeleteManyInstances(runtimeToken, instanceIds);
 
         //Find more instances to loop through
-        res = sbl.searchSblInstances(runtimeToken, filters);
-        success = check(res, {
-          'GET SBL Instance by Party status is 200': (r) => r.status === 200,
-        });
-        addErrorCount(success);
-
-        instances = sbl.filterInstancesByAppName(appNames, res.body);
-        instancesCount = instances.length;
+        res = instances.getAllinstancesWithFilters(runtimeToken, filters);
+        resBody = JSON.parse(res.body);
+        instanceIds = resBody.instances.map((i) => i.id);
+        instancesCount = instanceIds.count;
       }
     } while (instancesCount > 0);
+
+    success = check(instancesCount, {
+      'Hard delete instances for party. Remaining instance count is 0': (c) => c === 0,
+    });
+    addErrorCount(success);
   });
 }
